% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webpea.R
\name{webpea}
\alias{webpea}
\title{webpea}
\usage{
webpea(filename, path_return = TRUE, quality = NULL, ggsave = TRUE, ...)
}
\arguments{
\item{filename}{Filename to save the webp image to.}

\item{path_return}{Boolean, whether the absolute output path should be returned.
This allows to wrap the function in another function that works with the returned image. See
vignette for details. By default true.}

\item{quality}{Integer between 1 and 100. Specifies output quality for webp format. See {magick}
documentation for details. Defaults to reasonable 75, when not given.}

\item{ggsave}{Not yet in use.}

\item{...}{Arguments to be passed to ggplot::ggsave()}
}
\value{
Filename or path to the saved webp image.
}
\description{
Saves graphs in *.webp image format.
}
\examples{

# save last plot with default ggsave settings

ggplot2::ggplot(mtcars) +
  ggplot2::aes(disp, hp, color = cyl) +
  ggplot2::geom_point(alpha = 0.1)

webpea(tempfile("plot", fileext = ".webp"))

# save specific plot with optional ggsave paramaters and in higher-than-default quality

p1 <- ggplot2::ggplot(mtcars, ) +
  ggplot2::aes(disp, hp, color = cyl) +
  ggplot2::geom_point(alpha = 0.4)

webpea(
  tempfile("plot", fileext = ".webp"),
  plot = p1,
  width = 16,
  height = 9,
  quality = 90
)

}
